+++ .changes/unreleased/New feature-20230617-041635.yaml
@@ -1,0 +1,6 @@
+kind: New feature
+body: Global hooks
+time: 2023-06-17T04:16:35.435962+04:00
+custom:
+  Author: zhilyaev
+  Issue: "412"
