+++ .changes/unreleased/New feature-20230617-043909.yaml
@@ -1,0 +1,6 @@
+kind: New feature
+body: short aliases for hook scripting
+time: 2023-06-17T04:39:09.1662+04:00
+custom:
+  Author: zhilyaev
+  Issue: ""
