+++ pkg/hooks/config.go
@@ -7,26 +7,6 @@ import (
 	log "github.com/sirupsen/logrus"
 )
 
-type Global struct {
-	PreBuild  []hook `yaml:"pre_build"`
-	PostBuild []hook `yaml:"post_build"`
-
-	PreUp  []hook `yaml:"pre_up"`
-	PostUp []hook `yaml:"post_up"`
-
-	PreRollback  []hook `yaml:"pre_rollback"`
-	PostRollback []hook `yaml:"post_rollback"`
-
-	PreDown  []hook `yaml:"pre_down"`
-	PostDown []hook `yaml:"post_down"`
-}
-
-func Run(hooks []hook) {
-	for _, h := range hooks {
-		h.Run()
-	}
-}
-
 type Hook interface {
 	Run()
 	Log() *log.Entry
